+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.385.0 - 2026-06-16
 
-* `Aws\S3` - Added support for copying tags and annotations to the destination object in `MultipartCopy`. Set `copy_props` to `default` to copy metadata, tags, and annotations. Set `tags_directive` and `annotations_directive` to override individually. Tag and annotation work runs only when explicitly opted in to preserve backwards compatibility.
+* `Aws\S3` - Added support for copying tags and annotations to the destination object in `MultipartCopy`. Set `tags_directive` and `annotations_directive` to override individually. Tag and annotation work runs only when explicitly opted in to preserve backwards compatibility.
 
 ## 3.384.11 - 2026-06-16